chore(EMS-3876): update shared e2e currency assertions to test save/back functionality
