+++ e2e-tests/shared-test-assertions/currency-form-fields/form-submission/select-alt-radio-but-no-alt-currency.js
@@ -1,0 +1,9 @@
+const selectAltRadioButNoAltCurrency = ({ errorIndex = 0, rendersAlternativeCurrencyValidationError }) => {
+  describe('when selecting the alternative currency radio and NOT entering an alternative currency via the autocomplete input', () => {
+    it('should pre-select the alternative currency option and render validation errors', () => {
+      rendersAlternativeCurrencyValidationError({ errorIndex });
+    });
+  });
+};
+
+export default selectAltRadioButNoAltCurrency;
